Add files via upload
Excel file with The Fiery Passion Job Search Game - a game to motivate you to search for jobs in a way that's effective, not just launching applications into the void (even though that's a quicker, easier dopamine fix in the short run). Designed especially with neurodiverse, ADHDers in mind. 
Pick your rewards, track your job search activity, and then treat yourself for the process, not just the result!